docs: Update OverrideMistake.java

+++ src/IT독서/열혈자바프로그래밍/클래스의상속3_상속의목적/Object클래스와final선언그리고Override_2/OverrideMistake.java
@@ -8,7 +8,7 @@ class ParentAdder{
 
 class ChildAdder extends ParentAdder {
     @Override
-    public double add(double a, double b){
+    public int add(int a, int b){
         System.out.println("덧셈을 진행합니다.");
         return a+b;
     }
